\name{rankSVMcompare-package}
\alias{rankSVMcompare-package}
\alias{rankSVMcompare}
\docType{package}
\title{Support vector machines for ranking and comparing}
\description{Solvers for ranking and comparing problems,
where training data are paired features Xi and Xip,
with labels yi in c(-1,1) for ranking and c(-1,0,1) for comparing.}
\details{
\tabular{ll}{Package: \tab rankSVMcompare\cr
Maintainer: \tab Toby Dylan Hocking <toby@sg.cs.titech.ac.jp>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 2013.9.3\cr
License: \tab GPL-3\cr
Title: \tab Support vector machines for ranking and comparing\cr
Suggests: \tab ggplot2, directlabels, quadmod\cr
Depends: \tab kernlab\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


